Split validation schema and constants from main file
