+++ src/Components/payment-form/paymentForm.constant.js
@@ -1,0 +1,6 @@
+export const ValidLengths = {
+  MIN_CARDHOLDER_NAME_LENGTH: 2,
+  CARD_NUMBER_LENGTH: 16,
+  SECURITY_NUMBER_LENGTH: 3,
+  ACCOUNT_NUMBER_LENGTH: 10,
+};
